requirements.txt
django<2
django-polymorphic
